+++ tests/testthat/test-list-reporter/test-only-error.R
@@ -1,0 +1,1 @@
+stop('dying outside of tests')